requirements.txt
pycrypto==2.6.1
requests==2.13.0